# spring-boot-starter-template/src/main/java/fr/dlesaout/springbootstartertemplate/entity/Pet.java
package fr.dlesaout.springbootstartertemplate.entity;

import jakarta.persistence.*;
import lombok.*;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@EqualsAndHashCode
@ToString
@Builder
@Entity
@Table(name = "pets")
public class Pet {

    @Id
    @Column(name = "id")
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Integer id;

    @Column(name = "name")
    private String name;

    @Column(name = "tag")
    private String tag;

}